forensics/hostile-takeover/src/c2/requirements.txt
netifaces
flask
readline
pycryptodome
